--- old/example/index.html
@@ -1,85 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Scroll Snooper v1.2.2</title>
-    <link rel="icon" type="image/png" href="img/logo.png"/>
-    <link rel="stylesheet" href="styles.css">
-</head>
-<body>
-
-<header>
-    <div>
-        <img class="logo" src="img/logo.png" alt="Scroll Snooper Logo">
-        <h1>Scroll Snooper v1.2.2</h1>
-        <p>Pure JavaScript API that goes snooping around elements while scrolling</p>
-        <p>
-            <a href="https://github.com/phucbm/scroll-snooper/releases/latest">
-                <img src="https://badgen.net/github/release/phucbm/scroll-snooper/?cache=600" alt="release">
-            </a>
-            <a href="https://github.com/phucbm/scroll-snooper/blob/main/scroll-snooper.min.js">
-                <img src="https://badgen.net/badge/minified/4KB/cyan" alt="minified">
-            </a>
-            <a href="https://github.com/phucbm/scroll-snooper/blob/main/LICENSE">
-                <img src="https://badgen.net/github/license/phucbm/scroll-snooper/" alt="license">
-            </a>
-            <a href="https://www.jsdelivr.com/package/gh/phucbm/scroll-snooper">
-                <img src="https://data.jsdelivr.com/v1/package/gh/phucbm/scroll-snooper/badge?style=rounded"
-                     alt="jsdelivr">
-            </a>
-            <a href="https://github.com/phucbm/scroll-snooper">
-                <img src="https://badgen.net/badge/icon/github?icon=github&label" alt="jsdelivr">
-            </a>
-        </p>
-        <p class="giphy">
-            <img src="img/down-kelly-rowland-kisses-low.webp" alt="down-kelly-rowland-kisses-low">
-            <a href="https://giphy.com/gifs/down-kelly-rowland-kisses-low-LkuPxRS0F6gmc">via GIPHY</a>
-        </p>
-    </div>
-</header>
-
-<section class="scene-1">
-    <div class="bg"><i></i></div>
-    <div class="sticky">
-        <h2>CSS manipulating on scroll.</h2>
-    </div>
-    <div class="inner">
-        <img class="spinning-face" src="img/spinning-face.webp" alt="spinning-face">
-    </div>
-</section>
-
-<section class="scene-2">
-    <div class="sticky">
-        <h2>It's all about data.</h2>
-        <h4>Progress: <span data-progress></span></h4>
-        <h4>Visible pixel: <span data-pixel></span></h4>
-        <h4>Visible proportion: <span data-proportion></span></h4>
-        <p class="events">
-            <span class="event onScroll">onScroll</span>
-            <span class="event onEnter">onEnter</span>
-            <span class="event onLeave">onLeave</span>
-        </p>
-    </div>
-    <img class="shark" src="img/shark.webp" alt="shark">
-</section>
-
-<section class="scene-3">
-    <div class="sticky">
-        <h2>Custom start/end position (inspired by Scroll Trigger)</h2>
-        <h4>start: 'top 80%' , end: 'bottom center'</h4>
-        <h4>Progress: <span data-progress></span></h4>
-    </div>
-
-    <div class="image phoebe">
-        <img src="img/phoebe.webp" alt="phoebe">
-    </div>
-
-</section>
-
-
-<script src="https://cdn.jsdelivr.net/gh/phucbm/scroll-snooper@latest/scroll-snooper.min.js"></script>
-<script src="script.js"></script>
-
-</body>
-</html>
