--- old/package.json
@@ -1,10 +1,0 @@
-{
-  "devDependencies": {
-    "browser-sync": "^2.27.7",
-    "gulp": "^4.0.2",
-    "gulp-rename": "^2.0.0",
-    "gulp-replace": "^1.1.3",
-    "gulp-uglify": "^3.0.2",
-    "readline-sync": "^1.4.10"
-  }
-}
