+++ web/html/home.html
@@ -1,7 +1,68 @@
-<div class="container">
 
-    <h1>Scroll Snooper</h1>
+<header>
+    <div>
+        <img class="logo" src="../../public/img/logo.png" alt="Scroll Snooper Logo">
+        <h1>Scroll Snooper v1.2.2</h1>
+        <p>Pure JavaScript API that goes snooping around elements while scrolling</p>
+        <p>
+            <a href="https://github.com/phucbm/scroll-snooper/releases/latest">
+                <img src="https://badgen.net/github/release/phucbm/scroll-snooper/?cache=600" alt="release">
+            </a>
+            <a href="https://github.com/phucbm/scroll-snooper/blob/main/scroll-snooper.min.js">
+                <img src="https://badgen.net/badge/minified/4KB/cyan" alt="minified">
+            </a>
+            <a href="https://github.com/phucbm/scroll-snooper/blob/main/LICENSE">
+                <img src="https://badgen.net/github/license/phucbm/scroll-snooper/" alt="license">
+            </a>
+            <a href="https://www.jsdelivr.com/package/gh/phucbm/scroll-snooper">
+                <img src="https://data.jsdelivr.com/v1/package/gh/phucbm/scroll-snooper/badge?style=rounded"
+                     alt="jsdelivr">
+            </a>
+            <a href="https://github.com/phucbm/scroll-snooper">
+                <img src="https://badgen.net/badge/icon/github?icon=github&label" alt="jsdelivr">
+            </a>
+        </p>
+        <p class="giphy">
+            <img src="../../public/img/down-kelly-rowland-kisses-low.webp" alt="down-kelly-rowland-kisses-low">
+            <a href="https://giphy.com/gifs/down-kelly-rowland-kisses-low-LkuPxRS0F6gmc">via GIPHY</a>
+        </p>
+    </div>
+</header>
 
-    <div data-scroll-snooper></div>
+<section class="scene-1">
+    <div class="bg"><i></i></div>
+    <div class="sticky">
+        <h2>CSS manipulating on scroll.</h2>
+    </div>
+    <div class="inner">
+        <img class="spinning-face" src="../../public/img/spinning-face.webp" alt="spinning-face">
+    </div>
+</section>
 
-</div>
+<section class="scene-2">
+    <div class="sticky">
+        <h2>It's all about data.</h2>
+        <h4>Progress: <span data-progress></span></h4>
+        <h4>Visible pixel: <span data-pixel></span></h4>
+        <h4>Visible proportion: <span data-proportion></span></h4>
+        <p class="events">
+            <span class="event onScroll">onScroll</span>
+            <span class="event onEnter">onEnter</span>
+            <span class="event onLeave">onLeave</span>
+        </p>
+    </div>
+    <img class="shark" src="../../public/img/shark.webp" alt="shark">
+</section>
+
+<section class="scene-3">
+    <div class="sticky">
+        <h2>Custom start/end position (inspired by Scroll Trigger)</h2>
+        <h4>start: 'top 80%' , end: 'bottom center'</h4>
+        <h4>Progress: <span data-progress></span></h4>
+    </div>
+
+    <div class="image phoebe">
+        <img src="../../public/img/phoebe.webp" alt="phoebe">
+    </div>
+
+</section>
