+++ web/styles/_scaffolding.scss
@@ -8,8 +8,8 @@
 }
 
 body {
-  font-family: 'Finlandica', sans-serif;
-  padding: 20px;
+  font-family: AppleSystemUIFont, sans-serif;
+  font-size: 16px;
   margin: 0;
 }
 
@@ -17,4 +17,154 @@ body {
   max-width: $page-width;
   margin-left: auto;
   margin-right: auto;
+}
+
+header {
+  text-align: center;
+  padding: 2vw 0 2vw;
+  height: 100vh;
+  display: grid;
+  place-items: center;
+}
+
+header img.logo {
+  width: 20vw;
+  max-width: 100px;
+}
+
+footer {
+  height: 100vh;
+  display: grid;
+  place-items: center;
+  background-color: #0d47a1;
+  color: #fff;
+  background: -webkit-linear-gradient(bottom, rgba(0, 0, 0, 0) 0%, rgb(25, 118, 210) 100%);
+  background: -o-linear-gradient(bottom, rgba(0, 0, 0, 0) 0%, rgb(25, 118, 210) 100%);
+  background: -ms-linear-gradient(bottom, rgba(0, 0, 0, 0) 0%, rgb(25, 118, 210) 100%);
+  background: -moz-linear-gradient(bottom, rgba(0, 0, 0, 0) 0%, rgb(25, 118, 210) 100%);
+  background: linear-gradient(to top, rgba(0, 0, 0, 0) 0%, rgb(25, 118, 210) 100%);
+}
+
+.giphy img {
+  display: block;
+  margin: 0 auto 3px;
+  width: 20vw;
+  max-width: 100px;
+  border-radius: 4px;
+}
+
+.giphy a {
+  font-size: .6em;
+}
+
+h1 {
+  font-size: 1.6em;
+}
+
+h4 {
+  font-size: 1.1em;
+  margin-block-start: 0.3em;
+  margin-block-end: 0.3em;
+}
+
+a {
+  text-decoration: none;
+}
+
+/* events */
+.event {
+  background: #fff;
+  color: #000;
+  padding: .2em .4em;
+  border-radius: .2em;
+}
+
+.event.on {
+  background: limegreen;
+}
+
+/* sticky */
+.sticky {
+  position: sticky;
+  top: 0;
+  z-index: 2;
+  text-align: center;
+  padding: 2vw 0;
+}
+
+/* Scene 1 */
+.scene-1 {
+  height: 200vh;
+  background-color: #64b5f6;
+  position: relative;
+  color: #fff;
+}
+
+.scene-1 .inner {
+  position: sticky;
+  top: 0;
+  height: 100vh;
+  display: grid;
+  place-items: center;
+}
+
+.spinning-face {
+  width: 20vw;
+  height: 20vw;
+}
+
+.scene-1 .bg, .scene-1 .bg i {
+  position: absolute;
+  top: 0;
+  left: 0;
+  bottom: 0;
+  right: 0;
+  overflow: hidden;
+}
+
+.scene-1 .bg i {
+  background-color: #fff;
+}
+
+/* Scene 2 */
+.scene-2 {
+  height: 250vh;
+  background-color: #42a5f5;
+  color: #fff;
+  position: relative;
+}
+
+.scene-2 .shark {
+  position: absolute;
+  top: 100vh;
+  left: 50%;
+  transform: translate(-50%, 0);
+  width: 70vw;
+  max-width: 480px;
+  border-radius: 1vw;
+}
+
+/* Scene 3 */
+.scene-3 {
+  background-color: #1976d2;
+  color: #fff;
+}
+
+.scene-3 .image {
+  width: 90vw;
+  max-width: 480px;
+  margin: 0 auto;
+  padding: 40vw 0;
+}
+
+.scene-3 .image img {
+  border-radius: 1vw;
+}
+
+body:not(.scene-3-show) > #root > div {
+  display: none;
+}
+
+body > #root > div {
+  background-color: rgba(13, 71, 161, 0.5) !important;
 }
