+++ web/styles/_variables.scss
@@ -1,5 +1,5 @@
 $font-size: 1rem;
 $font-family: 'Finlandica', sans-serif;
-$page-width: 768px;
+$page-width: 1440px;
 $color-primary: #f1c400;
 $color-on-primary: #141414;
