+++ web/template.html
@@ -6,6 +6,7 @@
     <meta http-equiv="x-ua-compatible" content="ie=edge"/>
 
     <title>Scroll Snooper</title>
+    <link rel="icon" type="image/png" href="../public/img/logo.png"/>
 </head>
 
 <body>
